pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-task-graph-manager"
version = "0.1.0"
description = "Local-first personal task graph manager"
requires-python = ">=3.10"
dependencies = ["PyYAML>=5.4"]

[project.optional-dependencies]
test = ["pytest>=7"]

[tool.setuptools.packages.find]
include = ["taskmgr*"]
